Permutation with arrays(D5 P3)
Given a number N and two arrays A, B of N numbers. Determine if B is a permutation of A or not.

Note: A permutation is an arrangement of all or part of a set of objects.

For example: The array [2, 1, 3], [3, 2, 1] and [2, 3, 1] are permutation of the array [1, 2, 3].

Input
First line contains a number N (1≤N≤103) Number of elements.

Second line contains N numbers (1≤Ai≤107) elements of array A.

Third line contains N numbers (1≤Bi≤107) elements of array B.

Output
Print "yes" if array B is a permutation of A otherwise, print "no" without quotations.

Link:https://codeforces.com/group/MWSDmqGsZm/contest/219774/problem/R