Nice Grid (D3 P4)
Problem Statement
Print the color of the cell at the R-th row from the top and C-th column from the left in the following grid with 15 vertical rows and 15 horizontal columns.

Input:

Input is given from Standard Input in the following format:

R C

Output
In the grid above, if the color of the cell at the R-th row from the top and C-th column from the left is black, then print black; if the cell is white, then print white. Note that the judge is case-sensitive.

Link:https://atcoder.jp/contests/abc264/tasks/abc264_b