Full House(D5 P4)
Problem Statement
We have five cards with integers A, B, C, D, and E written on them, one on each card.

This set of five cards is called a Full house if and only if the following condition is satisfied:

the set has three cards with a same number written on them, and two cards with another same number written on them.
Determine whether the set is a Full house.

Constraints
1≤A,B,C,D,E≤13
Not all of A, B, C, D, and E are the same.
All values in input are integers.
Input
Input is given from Standard Input in the following format:

A B C D E
Output
If the set is a Full house, print Yes; otherwise, print No.

Link:https://atcoder.jp/contests/abc263/tasks/abc263_a